+++ src/server-loader/src/ModuleLoader.lua
@@ -10,6 +10,7 @@ local extractFunctionName = Common.extractFunctionName
 local defaultCustomModuleFilter = Common.defaultCustomModuleFilter
 local defaultExcludeModuleFilter = Common.defaultExcludeModuleFilter
 local filterArray = Common.filterArray
+local noYield = Common.noYield
 local ServerRemotes = require('./ServerRemotes')
 
 type Reporter = Common.Reporter
@@ -134,7 +135,13 @@ function ModuleLoader:loadModules()
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profilebegin(moduleInfo.name .. '.Init')
             end
-            moduleInfo.module.Init()
+
+            if _G.DEV then
+                noYield(moduleInfo.module.Init)
+            else
+                moduleInfo.module.Init()
+            end
+
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profileend()
             end
@@ -148,7 +155,13 @@ function ModuleLoader:loadModules()
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profilebegin(moduleInfo.name .. '.Init')
             end
-            moduleInfo.module.Init()
+
+            if _G.DEV then
+                noYield(moduleInfo.module.Init)
+            else
+                moduleInfo.module.Init()
+            end
+
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profileend()
             end
@@ -165,7 +178,13 @@ function ModuleLoader:loadModules()
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profilebegin(moduleInfo.name .. '.Start')
             end
-            moduleInfo.module.Start()
+
+            if _G.DEV then
+                noYield(moduleInfo.module.Start)
+            else
+                moduleInfo.module.Start()
+            end
+
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profileend()
             end
@@ -179,7 +198,13 @@ function ModuleLoader:loadModules()
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profilebegin(moduleInfo.name .. '.Start')
             end
-            moduleInfo.module.Start()
+
+            if _G.DEV then
+                noYield(moduleInfo.module.Start)
+            else
+                moduleInfo.module.Start()
+            end
+
             if _G.DEV or _G.CROSSWALK_PROFILE then
                 debug.profileend()
             end